Move vmmhz_em inside Askaryan Freqs, create ShowerProperties type
vmmhz_em functionality is probably broken,  but it looks like a
purely debugging spectator type of variable so it could be
deleted. Currently it is generated for everything which may well just
waste time.

moved emfrac and hadfrac inside a mini ShowerProperties class. More
stuff could be added in here. The idea for ShowerProperties is to move
things between Secondaries (i.e. the interesting shower info) and the
AskaryanFrequencies class.