Added FTPair class, currently just time domain constructors.
This is a coupled version of time domain/frequency domain signals.
All the complicated  business of tracking the state and doing the FTs
is handled internally.
It also does the checking required for numerical recipe's realft
function, which only does FTs on arrays that are a power of 2. That
cost me a morning...