Somehow this was missed when AnitaAnalysisFramework had renamed files.
For the A4NoNotches TUFF-digitizer response files per antenna, renaming files was done around beginning of July, but the accessing of these files with their new names was missed in this line of code. The TUFF configuration here is no notches on, so is really not used except in the case of the trigger efficiency scan.